DAOS-2660 common: move property API to prop.c
common/misc.c is overloaded, property API implementation should
be put in a separate file.
